requirements.txt
amqp==2.5.1
app-version==1.0.1
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coverage==4.5.1
coveralls==1.5.1
dataclasses==0.6
Django==2.2.3
django-admin-interface==0.10.5
django-admin-list-filter-dropdown==1.0.3
django-appconf==1.0.3
django-audiofield==0.10.0
django-celery==3.3.1
django-ckeditor==5.7.1
django-colorfield==0.1.15
django-cors-headers==3.1.0
django-cors-middleware==1.2.0
django-debug-toolbar==1.10
django-extensions==1.7.1
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-permission==1.0.4
django-ratings==0.3.7
django-redis==4.10.0
django-smart-selects==1.5.4
djangoql==0.13.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
djongo==1.2.33
docopt==0.6.2
Dumper==1.2.0
env-file==2019.4.13
factory-boy==2.11.1
fake-gen==1.0.0b2
Faker==0.9.2
idna==2.7
importlib-metadata==0.19
jongos==1.0
kombu==4.6.4
Markdown==2.6.11
mongoengine==0.18.2
more-itertools==7.2.0
mysqlclient==1.4.2.post1
packaging==19.1
permission==0.4.1
Pillow==5.2.0
pipenv==2018.11.26
pluggy==0.12.0
ply==3.11
public==2019.4.13
py==1.8.0
pycryptodome==3.8.2
PyJWT==1.6.4
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.1.1
python-dateutil==2.7.3
python-dotenv==0.10.3
python-memcached==1.59
pytz==2018.5
redis==3.3.11
requests==2.19.1
six==1.10.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.23
values==2019.4.13
vine==1.3.0
virtualenv==16.7.3
virtualenv-clone==0.5.3
wcwidth==0.1.7
wincertstore==0.2
zipp==0.6.0
